+++ zhtp/src/runtime/services/transaction_builder.rs
@@ -53,7 +53,7 @@ impl TransactionBuilder {
         use lib_economy::transactions::creation::create_reward_transaction;
 
         // Create reward for network services (routing, storage, etc.)
-        let reward_tx = create_reward_transaction(node_id, reward_amount)?;
+        let reward_tx = create_reward_transaction(node_id, reward_amount as u128)?;
 
         // Convert economics transaction to blockchain transaction
         Self::convert_economics_to_system_tx(&reward_tx, environment).await
@@ -148,7 +148,7 @@ impl TransactionBuilder {
             transaction_type: tx_type,
             inputs: inputs.to_vec(),
             outputs: outputs.to_vec(),
-            fee: economics_tx.total_fee,
+            fee: economics_tx.total_fee as u64,
             signature: temp_signature,
             memo: format!(
                 "System TX: {} {} SOV from {:?} to {:?}",
